Update LICENSE
Modified LICENSE to include MSU Denver as a copyright holder for our customized version of this resume template. This is in line with open source best practices: when you significantly modify or redistribute MIT-licensed software, it’s appropriate to update the license with your own copyright notice while preserving the original. This shows respect for the original author and clarifies ownership/responsibility for the modified version.

Signed-off-by: Maciek <1977132+Xata@users.noreply.github.com>